pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "accessvision"
version = "0.1.0"
description = "AI-powered visual WCAG accessibility auditing"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40.0",
    "firecrawl-py>=0.0.16",
    "google-generativeai>=0.3.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
